Split up `x86_il.c`
    * Split `x86_il.c` into `common.c` for the utility functions and
      helpers and `il_ops.inc` for the IL ops implementation
    * Now the `x86_il.c` file only contains the instruction to IL op
      mapping and the IL plugin
    * This is needed for adding more instructions, which now can be
      easily added by creating a new ".inc" file
