+++ librz/analysis/meson.build
@@ -164,6 +164,7 @@ rz_analysis_sources = [
   'arch/ppc/ppc_il_ops.c',
   'arch/ppc/ppc_il_flag_ops.c',
   'arch/sh/sh_il.c',
+  'arch/x86/common.c',
   'arch/x86/x86_il.c',
   'arch/8051/8051_il.c',
   'arch/8051/8051_parse.c',
